requirements.txt
langchain_core==0.3.43
langchain_google_genai==2.0.11
python-dotenv==1.0.1
